+++ layers/GraphQLAPIForWP/phpunit-packages/graphql-api-for-wp/tests/Integration/AbstractDefaultSchemaConfigurationForEndpointQueryExecutionModifyPluginSettingsFixtureEndpointWebserverRequestTest.php
@@ -1,0 +1,28 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PHPUnitForGraphQLAPI\GraphQLAPI\Integration;
+
+use GraphQLAPI\GraphQLAPI\Constants\ModuleSettingOptions;
+
+abstract class AbstractDefaultSchemaConfigurationForEndpointQueryExecutionModifyPluginSettingsFixtureEndpointWebserverRequestTest extends AbstractModifyPluginSettingsFixtureEndpointWebserverRequestTestCase
+{
+    public const SCHEMA_CONFIGURATION_WEBSITE_ID = 191;
+
+    protected function getFixtureFolder(): string
+    {
+        return __DIR__ . '/fixture-default-schema-configuration';
+    }
+
+    protected function getSettingsKey(): string
+    {
+        return ModuleSettingOptions::SCHEMA_CONFIGURATION;
+    }
+
+    protected function getPluginSettingsNewValue(): mixed
+    {
+        // New value: Schema Config "Website"
+        return self::SCHEMA_CONFIGURATION_WEBSITE_ID;
+    }
+}
